102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,15 +10,15 @@
 
 int main(void)
 {
-unsigned int f1 = 0;
-unsigned int f2 = 1;
+unsigned long f1 = 0;
+unsigned long f2 = 1;
 int count;
 unsigned long sum;
 _putchar(' ');
 while (count < 50)
 {
 sum = f1 + f2;
-printf("%ld, ", sum);
+printf("%lu, ", sum);
 f1 = f2;
 f2 = sum;
 count++;
